created a view using bootstrap, transferred user data to a modal window, implemented editing, adding and deleting users

+++ src/main/java/ru/kata/spring/boot_security/demo/model/User.java
@@ -4,7 +4,6 @@ package ru.kata.spring.boot_security.demo.model;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
-import org.hibernate.validator.constraints.UniqueElements;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
@@ -35,7 +34,7 @@ public class User implements UserDetails {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
 
     @Column(unique = true)
     @NotEmpty(message = "Name should not be empty")
